refactor: move hover/focus color derivation into OptionRow.Avatar
